requirements.txt
pyTelegramBotAPI==3.6.6
apiai==1.2.3
requests==2.20.0
Django==2.2.6
djangorestframework==3.10.3
PyYAML==5.1.2
